+++ app/Providers/Filament/AdminPanelProvider.php
@@ -2,6 +2,7 @@
 
 namespace App\Providers\Filament;
 
+use App\Filament\Pages\ManageUmamiSettings;
 use Filament\Http\Middleware\Authenticate;
 use Filament\Http\Middleware\AuthenticateSession;
 use Filament\Http\Middleware\DisableBladeIconComponents;
@@ -47,7 +48,10 @@ class AdminPanelProvider extends PanelProvider
                 fn (): View => view('elements.footer'),
             )
             ->plugins([
-                UmamiPlugin::make(),
+                UmamiPlugin::make()->settingsPage(false),
+            ])
+            ->pages([
+                ManageUmamiSettings::class,
             ])
             ->middleware([
                 EncryptCookies::class,